infra: rename singleton-export thing modules
