+++ src/data/things/WikiInfo.js
