+++ src/data/things/index.js
@@ -1,20 +1,21 @@
 // Not actually the entry point for #things - that's init.js in this folder.
 
-export * from './additional-file.js';
-export * from './additional-name.js';
+export * from './AdditionalFile.js';
+export * from './AdditionalName.js';
+export * from './ArtTag.js';
+export * from './Artist.js';
+export * from './Artwork.js';
+export * from './Contribution.js';
+export * from './Language.js';
+export * from './MusicVideo.js';
+export * from './NewsEntry.js';
+export * from './StaticPage.js';
+export * from './Track.js';
+export * from './WikiInfo.js';
+
 export * from './album.js';
-export * from './art-tag.js';
-export * from './artist.js';
-export * from './artwork.js';
 export * from './content.js';
-export * from './contribution.js';
 export * from './flash.js';
 export * from './group.js';
 export * from './homepage-layout.js';
-export * from './language.js';
-export * from './music-video.js';
-export * from './news-entry.js';
 export * from './sorting-rule.js';
-export * from './static-page.js';
-export * from './track.js';
-export * from './wiki-info.js';
